feat(rt-sdk): structured API errors, base64 feature gate, infallible default constructor
- Add ApiError/ErrorKind models mirroring the server error type
- Integrate ApiError into client response handling with reqwest fallback
- Gate base64 encoding/decoding behind optional `base64` feature
- Parse base URL once at construction, store as Url
- Make NvisyRt::new() infallible

+++ crates/nvisy-rt-sdk/src/client/config.rs
@@ -105,8 +105,8 @@ impl NvisyRtBuilder {
     }
 
     /// Sets the timeout in seconds.
-    pub fn with_timeout_secs(self, secs: u64) -> Self {
-        self.with_timeout(Duration::from_secs(secs))
+    pub fn with_timeout_secs(self, secs: u16) -> Self {
+        self.with_timeout(Duration::from_secs(secs as u64))
     }
 
     /// Builds the Nvisy Runtime client.
